Movement of P1 Paddle Bugfix
The initialization of the p1 and p2 paddles was such that they were previously added as non-physjs objects and were not controlled by dirtyposition commands. This has been fixed. Moving the scene.simulate() command on line 221 to line 237 fixes issues where the paddles will glitch out and fall through the ground/pop in and out of existence.