P1 Camera now follows p1 and the walls have collision interactions with the p1 character
The code now has the game board coded as physjs objects, and the camera will now follow the player 1 character while looking at the p2 character. This could be modified further by adding a look at p2 command, but that causes weird distortions to the view that might be detrimental to gameplay.