Made floor invisible and ball white
I also fixed a bit of style issues / commented. Since the camera is attached to the paddle itself, when the paddle falls over, the camera gets fucked. So I made the floor invisible so when the camera is facing upward (cause the paddle fell over) the player isn't looking straight into a wall.